Make InAppBrowser work with embedded files, using system behavior
The plugin is changed to work with embedded files, using the system behavior. If one needs the system behavior, they need to pass "_system" as target when using the plugin. As the InAppBrowser for WP8 does not handle by itself the embedded pdf files, system behavior is called.

The plugin also is tested and working for wp7