CB-7876 change test target to avoid undesired redirects
The google.com test target would do automatic redirects to geography-specific
domains depending on where the client is, which leads to different behavior
depending where the tester is. This causes confusion, as the test results
may be different in different geographies. Use a test target that doesn't do
that, so the test results are independent of geography.