Fix for CB-10305
There are really 3 issues here.
1. When the inAppBrowser is started in portrait and rotates, the
background of the status bar doesn’t fill the whole area. (any iOS
version)
2. On iOS 7, if the inAppBrowser plugin is started on landscape, the
background of the status bar was vertical.
3. On iOS 8-9, on iPhone/iPod, the status bar disappears on landscape,
but the background doesn’t disappear. I’ve changed it to make the
status bar appear  on landscape too. (This wasn’t on CB-10305, but I
think it’s related)