CB-12132: (android & ios) implement hidenotclose feature
This adds a new option hidenotclose which can be passed to the
inappbrowser.

When set to yes the done button on iOS or the X button android will
hide the web view instead of closing it.

This allows the user to show/hide the web view multiple times without
having to reload the content and you can clean up when you know they
are done by calling close once.