fix: merge issues caused by UIWebView and WKWebView implementation decouple
We had a fix in CDVInAppBrowser.m that commented
the code that allowed the UserAgent to be modified
(RPD-2199).

Most of that class has been "exploded" into multiple
other classes. This code in particular can be found in
CDVUIInAppBrowser.m and CDVWKInAppBrowser.m.

This commit fixes a merge issue caused by this
and passed the RPD-2199 fix to the new affected
code.

References https://outsystemsrd.atlassian.net/browse/RNMT-3220