feat: use WKWebView engine when available
The plugin supports UIWebView and optionally
WKWebView. The latter is only available if a
WKWebView engine plugin such as the one
we use in MABS 6 is available in the code.

Due to this we must still use the UIWebView version
in MABS 3, 4 and 5 and choose between both in
MABS 6. But considering Apple will soon stop
accepting apps with UIWebView it does not make
much sense to add an option for the user to choose
which engine to use.

We are then taking advantage of an already calculated
property that contains whether the WKWebView engine
is available or not to decide which engine to use (which
is WKWebView if available).

The original option is therefore rendered useless.

References https://outsystemsrd.atlassian.net/browse/RNMT-3220