fix: incorrect user-agent in the first usage of open
The setUserAgent method is supposed to manage the
User-Agent used by every UIWebView and WKWebView
in an application. For some reason this is not working
with our User-Agent in the first call of open (only when
using a WKWebView).

It works with UIWebView and it works with a different
value of User-Agent, which points to a potential bug in
the WKWebView engine plugin (possibly not detecting
a change to the value because one is a prefix of the other).

To fix this we are using the customUserAgent property
of the WKWebView. Setting this makes every request
use the intended value.

References https://outsystemsrd.atlassian.net/browse/RNMT-3299